divide monos into separate submodules
